// apps/figma-plugin/src/shared/types/intermediate-format.ts
export type TSVGNode = {
  readonly type: 'SVG';
  readonly svgHash: string;
} & TDefaultShapeMixin;
// ...
export type TDefaultShapeMixin = TBaseNodeMixin & TLayoutMixin & TBlendMixin;

export type TRectangleCornerMixin = {
  readonly topLeftRadius: number;
  readonly topRightRadius: number;
  readonly bottomLeftRadius: number;
  readonly bottomRightRadius: number;
};

export type TBaseNodeMixin = {
  readonly id: string;
  readonly name: string;
};

export type TChildrenMixin = {
  readonly children: ReadonlyArray<TNode>;
};

export type TLayoutMixin = {
  readonly width: number;
  readonly height: number;
  readonly rotation: number;
  readonly transform: TTransform;
} & TVector;

export type TFillsMixin = {
  readonly fills: ReadonlyArray<TPaint>;
};

// ============================================================================
// Effects
// ============================================================================

export type TBlendMixin = {
  readonly blendMode: TBlendMode;
  readonly opacity: number;
  readonly isMask: boolean;
  readonly effects: ReadonlyArray<TEffect>;
};